+++ domain/group.go
@@ -8,10 +8,5 @@ import (
 type GroupUser struct {
 	ID       uuid.UUID // Group ID
 	Name     string    // Group name
-	Duration GroupDuration
-}
-
-type GroupDuration struct {
-	Since YearWithSemester
-	Until YearWithSemester
+	Duration YearWithSemesterDuration
 }
