+++ domain/project.go
@@ -1,16 +1,13 @@
 package domain
 
 import (
-	"time"
-
 	"github.com/gofrs/uuid"
 )
 
 type Project struct {
 	ID          uuid.UUID
 	Name        string
-	Since       time.Time
-	Until       time.Time
+	Duration    YearWithSemesterDuration
 	Description string
 	Link        string
 	Members     []*ProjectMember
@@ -20,16 +17,5 @@ type ProjectMember struct {
 	UserID   uuid.UUID
 	Name     string
 	RealName string
-	Since    time.Time
-	Until    time.Time
-}
-
-type ProjectDuration struct {
-	Since YearWithSemester
-	Until YearWithSemester
-}
-
-type YearWithSemester struct {
-	Year     uint
-	Semester uint
+	Duration YearWithSemesterDuration
 }
